+++ .github/workflows/preview.yml
@@ -20,7 +20,7 @@ name: preview
 on:
   pull_request:
     branches:
-      - main
+#      - main
     paths-ignore:
       - '**.adoc'
       - '**.md'
