+++ endpoints/citrus-ftp/src/test/java/com/consol/citrus/ftp/client/SftpClientTest.java
@@ -54,7 +54,7 @@ import static org.testng.Assert.assertTrue;
  * @author Christoph Deppisch
  * @since 2.7.5
  */
-public class SftpClientTestASDF extends AbstractFtpClientTest {
+public class SftpClientTest extends AbstractFtpClientTest {
 
     private SftpClient sftpClient;
     private SshServer sshServer;
