+++ endpoints/citrus-jms/src/test/java/com/consol/citrus/jms/integration/SyncJmsTopicCommunicationIT.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.integration;
+package com.consol.citrus.jms.integration;
 
 import com.consol.citrus.annotations.CitrusXmlTest;
 import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
@@ -24,8 +24,8 @@ import org.testng.annotations.Test;
  * @author Christoph Deppisch
  * @since 2008
  */
-public class ValidateSqlResultSetITASDF extends TestNGCitrusSpringSupport {
+public class SyncJmsTopicCommunicationIT extends TestNGCitrusSpringSupport {
     @Test
     @CitrusXmlTest
-    public void ValidateSqlResultSetIT() {}
+    public void SyncJmsTopicCommunicationIT() {}
 }
