+++ validation/citrus-validation-groovy/src/test/java/com/consol/citrus/integration/ValidateSqlResultSetIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2006-2018 the original author or authors.
+ * Copyright 2006-2010 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.consol.citrus.ftp.integration;
+package com.consol.citrus.integration;
 
 import com.consol.citrus.annotations.CitrusXmlTest;
 import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
@@ -22,11 +22,10 @@ import org.testng.annotations.Test;
 
 /**
  * @author Christoph Deppisch
- * @since 2.7.5
+ * @since 2008
  */
-@Test
-public class FtpFileActionITASDF extends TestNGCitrusSpringSupport {
-
-    @CitrusXmlTest(name = "FtpFileActionIT")
-    public void testFileAction() {}
+public class ValidateSqlResultSetIT extends TestNGCitrusSpringSupport {
+    @Test
+    @CitrusXmlTest
+    public void ValidateSqlResultSetIT() {}
 }
